some sort of working type test

+++ src/freqresp.jl
@@ -38,7 +38,7 @@ function _preprocess_for_freqresp(sys::StateSpace)
     if isempty(sys.A) # hessfact does not work for empty matrices
         return sys
     end
-    TT = promote_type(numeric_type(sys), Float64)
+    TT = promote_type(numeric_type(sys), Float32)
 
     A, B, C, D = sys.A, sys.B, sys.C, sys.D
     F = hessfact(A)
